Partitioned the data into 3 segemnts, Training, Validation, and Test. Will need to clean
Attempt at adding regression analysis using interger values that we have: Gross, IMDB Rating, and Year. Appears that the Gross value in correlation to the year and IMDB is VERY low. ~7.6% (using a log transform to reduce outlier impact) Without transform we see it around a 6.8% instead. We may need to look into what was originally taken out and add it back in and test against a smaller subsection.